remove stuff that's in react-email now, reorganize utils
