--- packages/preview-server/src/utils/caniemail/get-css-unit.ts
@@ -1,3 +1,0 @@
-export const getCssUnit = (title: string) => {
-  return title.endsWith(' unit') ? title.replace(' unit', '') : undefined;
-};
