--- packages/preview-server/src/utils/caniemail/get-element-attributes.ts
@@ -1,7 +1,0 @@
-export function getElementAttributes(title: string) {
-  if (title.endsWith(' attribute')) {
-    return [title.replace(' attribute', '')];
-  }
-
-  return [];
-}
